Move message type
