+++ mb/kafka/kafka.go
@@ -3,30 +3,22 @@ package kafka
 import (
 	"context"
 	"fmt"
-	"time"
 
 	"github.com/nrfta/go-outbox"
+	"github.com/nrfta/go-outbox/mb"
 
 	"github.com/twmb/franz-go/pkg/kgo"
 )
-
-type Message struct {
-	Key       []byte
-	Value     []byte
-	Headers   []kgo.RecordHeader
-	Timestamp time.Time
-	Topic     string
-}
 
 type kafkaMessageBroker struct {
 	client *kgo.Client
 }
 
-func New(client *kgo.Client) (outbox.MessageBroker[Message], error) {
+func New(client *kgo.Client) (outbox.MessageBroker[mb.Message], error) {
 	return &kafkaMessageBroker{client}, nil
 }
 
-func (mb kafkaMessageBroker) Send(ctx context.Context, msg Message) error {
+func (mb kafkaMessageBroker) Send(ctx context.Context, msg mb.Message) error {
 	record := &kgo.Record{
 		Key:       msg.Key,
 		Value:     msg.Value,
